+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/tag/dto/TagRequest.kt
@@ -1,0 +1,8 @@
+package com.wafflestudio.toyproject.memoWithTags.tag.dto
+
+sealed class TagRequest {
+    data class CreateTagRequest(
+        val name: String,
+        val color: String
+    ) : TagRequest()
+}
