+++ src/main/kotlin/com/wafflestudio/toyproject/memoWithTags/tag/dto/TagResponse.kt
@@ -1,0 +1,8 @@
+package com.wafflestudio.toyproject.memoWithTags.tag.dto
+
+sealed class TagResponse {
+    data class UpdateTagRequest(
+        val name: String,
+        val color: String
+    ) : TagResponse()
+}
